api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
crewai>=0.11.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb>=0.4.22
pydantic>=2.6.0 